fix(usecase): DeleteRotationUseCase

+++ src/http/controller/rotation/deleteRotationController.ts
@@ -1,21 +1,33 @@
+import { BadRequestError } from "@/error/badRequest.error";
 import { PrismaRotationRepository } from "@/repository/prisma/prismaRotationRepository";
+import { PrismaUserRepository } from "@/repository/prisma/prismaUserRepository";
 import { DeleteRotationUseCase } from "@/usecase/rotation/deleteRotationUseCase";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { z } from "zod";
 
 export class DeleteRotationController {
   async handle(req: FastifyRequest, res: FastifyReply){
+    if (!req.userAuth?.id) {
+      throw new BadRequestError("userId not informed");
+    }
+    const userId = req.userAuth.id
+
     const deleteRotationRequestParams = z.object({
       rotationId: z.string().uuid()
     })
     const {rotationId} = deleteRotationRequestParams.parse(req.params)
 
     const rotationRepository = new PrismaRotationRepository()
+    const userRepository = new PrismaUserRepository()
     const deleteRotationUseCase = new DeleteRotationUseCase(
-      rotationRepository
+      rotationRepository,
+      userRepository
     )
 
-    await deleteRotationUseCase.execute(rotationId)
+    await deleteRotationUseCase.execute({
+      rotationId,
+      userId
+    })
 
     return res.status(200).send()
   }
